Return life table data stored by package

Return life table data stored by package

Usage

package_lt(country, year, lt_extend = TRUE)

Arguments

country

[string]

The name of a country (for which data is available & stored in the package). Case-sensitive. Please use function hrqol_norms to see the list of permissible country names.

year

[integer]

A year (for which data is available & stored in the package).

lt_extend

[boolean] or [numeric]

Allows users to control whether/ the way in which assumptions are made about mortality rates among people older than 99, for whom data is not typically available.

If FALSE, no assumption is made, and the function assumes no people live beyond 99.

If TRUE (default), the function assumes that people can live up to 120 and calculates a mortality rate for the older ages by assuming that mortality rates increase year on year by a constant increment, which is set equal to the average rate of increase over the last 10 years for which data is available.

Alternatively, the user can specify their own increment, instead of allowing the function to calculate an increment automatically based on existing data. This is done by passing lt_extend a numeric value greater than or equal to 1 - for example, setting lt_extend to 1.05 means the function assumes mortality rates will increase by 5% with every additional year of age gained after the last year for which data is available.

Value

A data frame, containing life tables for the chosen country and year.

Examples

package_lt(country = "Romania", year = 2022)
#>        sex age          q
#> 1   female   0 0.00480196
#> 2     male   0 0.00557985
#> 3   female   1 0.00056791
#> 4     male   1 0.00054797
#> 5   female   2 0.00034874
#> 6     male   2 0.00039457
#> 7   female   3 0.00029756
#> 8     male   3 0.00036220
#> 9   female   4 0.00020969
#> 10    male   4 0.00021677
#> 11  female   5 0.00012278
#> 12    male   5 0.00019949
#> 13  female   6 0.00012753
#> 14    male   6 0.00023453
#> 15  female   7 0.00016060
#> 16    male   7 0.00020172
#> 17  female   8 0.00010565
#> 18    male   8 0.00014283
#> 19  female   9 0.00017283
#> 20    male   9 0.00019773
#> 21  female  10 0.00016311
#> 22    male  10 0.00011388
#> 23  female  11 0.00013773
#> 24    male  11 0.00017172
#> 25  female  12 0.00016654
#> 26    male  12 0.00022355
#> 27  female  13 0.00020069
#> 28    male  13 0.00018943
#> 29  female  14 0.00013817
#> 30    male  14 0.00033847
#> 31  female  15 0.00029765
#> 32    male  15 0.00036501
#> 33  female  16 0.00021214
#> 34    male  16 0.00037150
#> 35  female  17 0.00031083
#> 36    male  17 0.00064137
#> 37  female  18 0.00029726
#> 38    male  18 0.00080546
#> 39  female  19 0.00026690
#> 40    male  19 0.00078819
#> 41  female  20 0.00023922
#> 42    male  20 0.00075764
#> 43  female  21 0.00036669
#> 44    male  21 0.00094881
#> 45  female  22 0.00027665
#> 46    male  22 0.00070086
#> 47  female  23 0.00030551
#> 48    male  23 0.00092733
#> 49  female  24 0.00033893
#> 50    male  24 0.00085175
#> 51  female  25 0.00029218
#> 52    male  25 0.00088226
#> 53  female  26 0.00031391
#> 54    male  26 0.00090246
#> 55  female  27 0.00036957
#> 56    male  27 0.00095270
#> 57  female  28 0.00043807
#> 58    male  28 0.00104500
#> 59  female  29 0.00042898
#> 60    male  29 0.00097607
#> 61  female  30 0.00049680
#> 62    male  30 0.00109131
#> 63  female  31 0.00035210
#> 64    male  31 0.00131046
#> 65  female  32 0.00049932
#> 66    male  32 0.00126064
#> 67  female  33 0.00046288
#> 68    male  33 0.00136055
#> 69  female  34 0.00064413
#> 70    male  34 0.00135130
#> 71  female  35 0.00062920
#> 72    male  35 0.00136332
#> 73  female  36 0.00063240
#> 74    male  36 0.00181321
#> 75  female  37 0.00087204
#> 76    male  37 0.00177075
#> 77  female  38 0.00087460
#> 78    male  38 0.00206263
#> 79  female  39 0.00092709
#> 80    male  39 0.00223435
#> 81  female  40 0.00094927
#> 82    male  40 0.00275506
#> 83  female  41 0.00120231
#> 84    male  41 0.00291491
#> 85  female  42 0.00145598
#> 86    male  42 0.00321591
#> 87  female  43 0.00146382
#> 88    male  43 0.00371947
#> 89  female  44 0.00156330
#> 90    male  44 0.00431794
#> 91  female  45 0.00176717
#> 92    male  45 0.00473702
#> 93  female  46 0.00196363
#> 94    male  46 0.00504245
#> 95  female  47 0.00244199
#> 96    male  47 0.00573333
#> 97  female  48 0.00268071
#> 98    male  48 0.00678773
#> 99  female  49 0.00284441
#> 100   male  49 0.00728255
#> 101 female  50 0.00334682
#> 102   male  50 0.00827357
#> 103 female  51 0.00386900
#> 104   male  51 0.00986142
#> 105 female  52 0.00352595
#> 106   male  52 0.00894093
#> 107 female  53 0.00423439
#> 108   male  53 0.01130906
#> 109 female  54 0.00449266
#> 110   male  54 0.01213933
#> 111 female  55 0.00514210
#> 112   male  55 0.01405176
#> 113 female  56 0.00542086
#> 114   male  56 0.01509369
#> 115 female  57 0.00658565
#> 116   male  57 0.01659987
#> 117 female  58 0.00711610
#> 118   male  58 0.01825062
#> 119 female  59 0.00760623
#> 120   male  59 0.01975647
#> 121 female  60 0.00843290
#> 122   male  60 0.02236177
#> 123 female  61 0.00922815
#> 124   male  61 0.02361570
#> 125 female  62 0.00964544
#> 126   male  62 0.02474014
#> 127 female  63 0.01002079
#> 128   male  63 0.02596069
#> 129 female  64 0.01208029
#> 130   male  64 0.02803118
#> 131 female  65 0.01205207
#> 132   male  65 0.02909234
#> 133 female  66 0.01328786
#> 134   male  66 0.03201797
#> 135 female  67 0.01423362
#> 136   male  67 0.03307030
#> 137 female  68 0.01566448
#> 138   male  68 0.03552632
#> 139 female  69 0.01754723
#> 140   male  69 0.03769957
#> 141 female  70 0.01909690
#> 142   male  70 0.03966633
#> 143 female  71 0.02033009
#> 144   male  71 0.04107671
#> 145 female  72 0.02300676
#> 146   male  72 0.04451605
#> 147 female  73 0.02621647
#> 148   male  73 0.04756734
#> 149 female  74 0.02876160
#> 150   male  74 0.05185886
#> 151 female  75 0.03375342
#> 152   male  75 0.05548635
#> 153 female  76 0.03721405
#> 154   male  76 0.06050392
#> 155 female  77 0.04161291
#> 156   male  77 0.06701584
#> 157 female  78 0.04450472
#> 158   male  78 0.06951167
#> 159 female  79 0.05318607
#> 160   male  79 0.07872881
#> 161 female  80 0.05942003
#> 162   male  80 0.08542466
#> 163 female  81 0.06617262
#> 164   male  81 0.09060592
#> 165 female  82 0.07778366
#> 166   male  82 0.09904959
#> 167 female  83 0.08429272
#> 168   male  83 0.10733591
#> 169 female  84 0.09834761
#> 170   male  84 0.11890722
#> 171 female  85 0.11071871
#> 172   male  85 0.12849540
#> 173 female  86 0.12552352
#> 174   male  86 0.14013393
#> 175 female  87 0.14229964
#> 176   male  87 0.15291446
#> 177 female  88 0.16131107
#> 178   male  88 0.16694964
#> 179 female  89 0.18280135
#> 180   male  89 0.18233473
#> 181 female  90 0.20696686
#> 182   male  90 0.19915223
#> 183 female  91 0.23400530
#> 184   male  91 0.21748540
#> 185 female  92 0.26410991
#> 186   male  92 0.23741709
#> 187 female  93 0.29751787
#> 188   male  93 0.25906118
#> 189 female  94 0.33447482
#> 190   male  94 0.28254581
#> 191 female  95 0.35915485
#> 192   male  95 0.30114462
#> 193 female  96 0.38998118
#> 194   male  96 0.32282437
#> 195 female  97 0.42028889
#> 196   male  97 0.34482557
#> 197 female  98 0.45009962
#> 198   male  98 0.36725149
#> 199 female  99 0.47929029
#> 200   male  99 0.39008160
#> 201 female 100 0.51267210
#> 202   male 100 0.41347251
#> 203 female 101 0.54697166
#> 204   male 101 0.43770820
#> 205 female 102 0.58196625
#> 206   male 102 0.46273860
#> 207 female 103 0.61739341
#> 208   male 103 0.48850030
#> 209 female 104 0.65295262
#> 210   male 104 0.51491564
#> 211 female 105 0.68830929
#> 212   male 105 0.54189206
#> 213 female 106 0.72310125
#> 214   male 106 0.56932172
#> 215 female 107 0.75694803
#> 216   male 107 0.59708145
#> 217 female 108 0.78946302
#> 218   male 108 0.62503307
#> 219 female 109 0.82026814
#> 220   male 109 0.65302430
#> 221 female 110 0.84901069
#> 222   male 110 0.68089013
#> 223 female 111 0.87538120
#> 224   male 111 0.70845493
#> 225 female 112 0.89913116
#> 226   male 112 0.73553512
#> 227 female 113 0.92008888
#> 228   male 113 0.76194258
#> 229 female 114 0.93817167
#> 230   male 114 0.78748870
#> 231 female 115 0.95339256
#> 232   male 115 0.81198908
#> 233 female 116 0.96586013
#> 234   male 116 0.83526856
#> 235 female 117 0.97577064
#> 236   male 117 0.85716669
#> 237 female 118 0.98339258
#> 238   male 118 0.87754318
#> 239 female 119 0.98904513
#> 240   male 119 0.89628313
#> 241 female 120 0.99307267
#> 242   male 120 0.91330169
package_lt(country = "Romania", year = 2022, lt_extend = FALSE)
#>        sex age          q
#> 1   female   0 0.00480196
#> 2     male   0 0.00557985
#> 3   female   1 0.00056791
#> 4     male   1 0.00054797
#> 5   female   2 0.00034874
#> 6     male   2 0.00039457
#> 7   female   3 0.00029756
#> 8     male   3 0.00036220
#> 9   female   4 0.00020969
#> 10    male   4 0.00021677
#> 11  female   5 0.00012278
#> 12    male   5 0.00019949
#> 13  female   6 0.00012753
#> 14    male   6 0.00023453
#> 15  female   7 0.00016060
#> 16    male   7 0.00020172
#> 17  female   8 0.00010565
#> 18    male   8 0.00014283
#> 19  female   9 0.00017283
#> 20    male   9 0.00019773
#> 21  female  10 0.00016311
#> 22    male  10 0.00011388
#> 23  female  11 0.00013773
#> 24    male  11 0.00017172
#> 25  female  12 0.00016654
#> 26    male  12 0.00022355
#> 27  female  13 0.00020069
#> 28    male  13 0.00018943
#> 29  female  14 0.00013817
#> 30    male  14 0.00033847
#> 31  female  15 0.00029765
#> 32    male  15 0.00036501
#> 33  female  16 0.00021214
#> 34    male  16 0.00037150
#> 35  female  17 0.00031083
#> 36    male  17 0.00064137
#> 37  female  18 0.00029726
#> 38    male  18 0.00080546
#> 39  female  19 0.00026690
#> 40    male  19 0.00078819
#> 41  female  20 0.00023922
#> 42    male  20 0.00075764
#> 43  female  21 0.00036669
#> 44    male  21 0.00094881
#> 45  female  22 0.00027665
#> 46    male  22 0.00070086
#> 47  female  23 0.00030551
#> 48    male  23 0.00092733
#> 49  female  24 0.00033893
#> 50    male  24 0.00085175
#> 51  female  25 0.00029218
#> 52    male  25 0.00088226
#> 53  female  26 0.00031391
#> 54    male  26 0.00090246
#> 55  female  27 0.00036957
#> 56    male  27 0.00095270
#> 57  female  28 0.00043807
#> 58    male  28 0.00104500
#> 59  female  29 0.00042898
#> 60    male  29 0.00097607
#> 61  female  30 0.00049680
#> 62    male  30 0.00109131
#> 63  female  31 0.00035210
#> 64    male  31 0.00131046
#> 65  female  32 0.00049932
#> 66    male  32 0.00126064
#> 67  female  33 0.00046288
#> 68    male  33 0.00136055
#> 69  female  34 0.00064413
#> 70    male  34 0.00135130
#> 71  female  35 0.00062920
#> 72    male  35 0.00136332
#> 73  female  36 0.00063240
#> 74    male  36 0.00181321
#> 75  female  37 0.00087204
#> 76    male  37 0.00177075
#> 77  female  38 0.00087460
#> 78    male  38 0.00206263
#> 79  female  39 0.00092709
#> 80    male  39 0.00223435
#> 81  female  40 0.00094927
#> 82    male  40 0.00275506
#> 83  female  41 0.00120231
#> 84    male  41 0.00291491
#> 85  female  42 0.00145598
#> 86    male  42 0.00321591
#> 87  female  43 0.00146382
#> 88    male  43 0.00371947
#> 89  female  44 0.00156330
#> 90    male  44 0.00431794
#> 91  female  45 0.00176717
#> 92    male  45 0.00473702
#> 93  female  46 0.00196363
#> 94    male  46 0.00504245
#> 95  female  47 0.00244199
#> 96    male  47 0.00573333
#> 97  female  48 0.00268071
#> 98    male  48 0.00678773
#> 99  female  49 0.00284441
#> 100   male  49 0.00728255
#> 101 female  50 0.00334682
#> 102   male  50 0.00827357
#> 103 female  51 0.00386900
#> 104   male  51 0.00986142
#> 105 female  52 0.00352595
#> 106   male  52 0.00894093
#> 107 female  53 0.00423439
#> 108   male  53 0.01130906
#> 109 female  54 0.00449266
#> 110   male  54 0.01213933
#> 111 female  55 0.00514210
#> 112   male  55 0.01405176
#> 113 female  56 0.00542086
#> 114   male  56 0.01509369
#> 115 female  57 0.00658565
#> 116   male  57 0.01659987
#> 117 female  58 0.00711610
#> 118   male  58 0.01825062
#> 119 female  59 0.00760623
#> 120   male  59 0.01975647
#> 121 female  60 0.00843290
#> 122   male  60 0.02236177
#> 123 female  61 0.00922815
#> 124   male  61 0.02361570
#> 125 female  62 0.00964544
#> 126   male  62 0.02474014
#> 127 female  63 0.01002079
#> 128   male  63 0.02596069
#> 129 female  64 0.01208029
#> 130   male  64 0.02803118
#> 131 female  65 0.01205207
#> 132   male  65 0.02909234
#> 133 female  66 0.01328786
#> 134   male  66 0.03201797
#> 135 female  67 0.01423362
#> 136   male  67 0.03307030
#> 137 female  68 0.01566448
#> 138   male  68 0.03552632
#> 139 female  69 0.01754723
#> 140   male  69 0.03769957
#> 141 female  70 0.01909690
#> 142   male  70 0.03966633
#> 143 female  71 0.02033009
#> 144   male  71 0.04107671
#> 145 female  72 0.02300676
#> 146   male  72 0.04451605
#> 147 female  73 0.02621647
#> 148   male  73 0.04756734
#> 149 female  74 0.02876160
#> 150   male  74 0.05185886
#> 151 female  75 0.03375342
#> 152   male  75 0.05548635
#> 153 female  76 0.03721405
#> 154   male  76 0.06050392
#> 155 female  77 0.04161291
#> 156   male  77 0.06701584
#> 157 female  78 0.04450472
#> 158   male  78 0.06951167
#> 159 female  79 0.05318607
#> 160   male  79 0.07872881
#> 161 female  80 0.05942003
#> 162   male  80 0.08542466
#> 163 female  81 0.06617262
#> 164   male  81 0.09060592
#> 165 female  82 0.07778366
#> 166   male  82 0.09904959
#> 167 female  83 0.08429272
#> 168   male  83 0.10733591
#> 169 female  84 0.09834761
#> 170   male  84 0.11890722
#> 171 female  85 0.11071871
#> 172   male  85 0.12849540
#> 173 female  86 0.12552352
#> 174   male  86 0.14013393
#> 175 female  87 0.14229964
#> 176   male  87 0.15291446
#> 177 female  88 0.16131107
#> 178   male  88 0.16694964
#> 179 female  89 0.18280135
#> 180   male  89 0.18233473
#> 181 female  90 0.20696686
#> 182   male  90 0.19915223
#> 183 female  91 0.23400530
#> 184   male  91 0.21748540
#> 185 female  92 0.26410991
#> 186   male  92 0.23741709
#> 187 female  93 0.29751787
#> 188   male  93 0.25906118
#> 189 female  94 0.33447482
#> 190   male  94 0.28254581
#> 191 female  95 0.35915485
#> 192   male  95 0.30114462
#> 193 female  96 0.38998118
#> 194   male  96 0.32282437
#> 195 female  97 0.42028889
#> 196   male  97 0.34482557
#> 197 female  98 0.45009962
#> 198   male  98 0.36725149
#> 199 female  99 0.47929029
#> 200   male  99 0.39008160
package_lt(country = "Romania", year = 2022, lt_extend = 1.5)
#>        sex age          q
#> 1   female   0 0.00480196
#> 2     male   0 0.00557985
#> 3   female   1 0.00056791
#> 4     male   1 0.00054797
#> 5   female   2 0.00034874
#> 6     male   2 0.00039457
#> 7   female   3 0.00029756
#> 8     male   3 0.00036220
#> 9   female   4 0.00020969
#> 10    male   4 0.00021677
#> 11  female   5 0.00012278
#> 12    male   5 0.00019949
#> 13  female   6 0.00012753
#> 14    male   6 0.00023453
#> 15  female   7 0.00016060
#> 16    male   7 0.00020172
#> 17  female   8 0.00010565
#> 18    male   8 0.00014283
#> 19  female   9 0.00017283
#> 20    male   9 0.00019773
#> 21  female  10 0.00016311
#> 22    male  10 0.00011388
#> 23  female  11 0.00013773
#> 24    male  11 0.00017172
#> 25  female  12 0.00016654
#> 26    male  12 0.00022355
#> 27  female  13 0.00020069
#> 28    male  13 0.00018943
#> 29  female  14 0.00013817
#> 30    male  14 0.00033847
#> 31  female  15 0.00029765
#> 32    male  15 0.00036501
#> 33  female  16 0.00021214
#> 34    male  16 0.00037150
#> 35  female  17 0.00031083
#> 36    male  17 0.00064137
#> 37  female  18 0.00029726
#> 38    male  18 0.00080546
#> 39  female  19 0.00026690
#> 40    male  19 0.00078819
#> 41  female  20 0.00023922
#> 42    male  20 0.00075764
#> 43  female  21 0.00036669
#> 44    male  21 0.00094881
#> 45  female  22 0.00027665
#> 46    male  22 0.00070086
#> 47  female  23 0.00030551
#> 48    male  23 0.00092733
#> 49  female  24 0.00033893
#> 50    male  24 0.00085175
#> 51  female  25 0.00029218
#> 52    male  25 0.00088226
#> 53  female  26 0.00031391
#> 54    male  26 0.00090246
#> 55  female  27 0.00036957
#> 56    male  27 0.00095270
#> 57  female  28 0.00043807
#> 58    male  28 0.00104500
#> 59  female  29 0.00042898
#> 60    male  29 0.00097607
#> 61  female  30 0.00049680
#> 62    male  30 0.00109131
#> 63  female  31 0.00035210
#> 64    male  31 0.00131046
#> 65  female  32 0.00049932
#> 66    male  32 0.00126064
#> 67  female  33 0.00046288
#> 68    male  33 0.00136055
#> 69  female  34 0.00064413
#> 70    male  34 0.00135130
#> 71  female  35 0.00062920
#> 72    male  35 0.00136332
#> 73  female  36 0.00063240
#> 74    male  36 0.00181321
#> 75  female  37 0.00087204
#> 76    male  37 0.00177075
#> 77  female  38 0.00087460
#> 78    male  38 0.00206263
#> 79  female  39 0.00092709
#> 80    male  39 0.00223435
#> 81  female  40 0.00094927
#> 82    male  40 0.00275506
#> 83  female  41 0.00120231
#> 84    male  41 0.00291491
#> 85  female  42 0.00145598
#> 86    male  42 0.00321591
#> 87  female  43 0.00146382
#> 88    male  43 0.00371947
#> 89  female  44 0.00156330
#> 90    male  44 0.00431794
#> 91  female  45 0.00176717
#> 92    male  45 0.00473702
#> 93  female  46 0.00196363
#> 94    male  46 0.00504245
#> 95  female  47 0.00244199
#> 96    male  47 0.00573333
#> 97  female  48 0.00268071
#> 98    male  48 0.00678773
#> 99  female  49 0.00284441
#> 100   male  49 0.00728255
#> 101 female  50 0.00334682
#> 102   male  50 0.00827357
#> 103 female  51 0.00386900
#> 104   male  51 0.00986142
#> 105 female  52 0.00352595
#> 106   male  52 0.00894093
#> 107 female  53 0.00423439
#> 108   male  53 0.01130906
#> 109 female  54 0.00449266
#> 110   male  54 0.01213933
#> 111 female  55 0.00514210
#> 112   male  55 0.01405176
#> 113 female  56 0.00542086
#> 114   male  56 0.01509369
#> 115 female  57 0.00658565
#> 116   male  57 0.01659987
#> 117 female  58 0.00711610
#> 118   male  58 0.01825062
#> 119 female  59 0.00760623
#> 120   male  59 0.01975647
#> 121 female  60 0.00843290
#> 122   male  60 0.02236177
#> 123 female  61 0.00922815
#> 124   male  61 0.02361570
#> 125 female  62 0.00964544
#> 126   male  62 0.02474014
#> 127 female  63 0.01002079
#> 128   male  63 0.02596069
#> 129 female  64 0.01208029
#> 130   male  64 0.02803118
#> 131 female  65 0.01205207
#> 132   male  65 0.02909234
#> 133 female  66 0.01328786
#> 134   male  66 0.03201797
#> 135 female  67 0.01423362
#> 136   male  67 0.03307030
#> 137 female  68 0.01566448
#> 138   male  68 0.03552632
#> 139 female  69 0.01754723
#> 140   male  69 0.03769957
#> 141 female  70 0.01909690
#> 142   male  70 0.03966633
#> 143 female  71 0.02033009
#> 144   male  71 0.04107671
#> 145 female  72 0.02300676
#> 146   male  72 0.04451605
#> 147 female  73 0.02621647
#> 148   male  73 0.04756734
#> 149 female  74 0.02876160
#> 150   male  74 0.05185886
#> 151 female  75 0.03375342
#> 152   male  75 0.05548635
#> 153 female  76 0.03721405
#> 154   male  76 0.06050392
#> 155 female  77 0.04161291
#> 156   male  77 0.06701584
#> 157 female  78 0.04450472
#> 158   male  78 0.06951167
#> 159 female  79 0.05318607
#> 160   male  79 0.07872881
#> 161 female  80 0.05942003
#> 162   male  80 0.08542466
#> 163 female  81 0.06617262
#> 164   male  81 0.09060592
#> 165 female  82 0.07778366
#> 166   male  82 0.09904959
#> 167 female  83 0.08429272
#> 168   male  83 0.10733591
#> 169 female  84 0.09834761
#> 170   male  84 0.11890722
#> 171 female  85 0.11071871
#> 172   male  85 0.12849540
#> 173 female  86 0.12552352
#> 174   male  86 0.14013393
#> 175 female  87 0.14229964
#> 176   male  87 0.15291446
#> 177 female  88 0.16131107
#> 178   male  88 0.16694964
#> 179 female  89 0.18280135
#> 180   male  89 0.18233473
#> 181 female  90 0.20696686
#> 182   male  90 0.19915223
#> 183 female  91 0.23400530
#> 184   male  91 0.21748540
#> 185 female  92 0.26410991
#> 186   male  92 0.23741709
#> 187 female  93 0.29751787
#> 188   male  93 0.25906118
#> 189 female  94 0.33447482
#> 190   male  94 0.28254581
#> 191 female  95 0.35915485
#> 192   male  95 0.30114462
#> 193 female  96 0.38998118
#> 194   male  96 0.32282437
#> 195 female  97 0.42028889
#> 196   male  97 0.34482557
#> 197 female  98 0.45009962
#> 198   male  98 0.36725149
#> 199 female  99 0.47929029
#> 200   male  99 0.39008160
#> 201 female 100 0.62425474
#> 202   male 100 0.52367036
#> 203 female 101 0.76967543
#> 204   male 101 0.67125330
#> 205 female 102 0.88946230
#> 206   male 102 0.81150835
#> 207 female 103 0.96324930
#> 208   male 103 0.91816517
#> 209 female 104 0.99295472
#> 210   male 104 0.97658969
#> 211 female 105 0.99940865
#> 212   male 105 0.99641812
#> 213 female 106 0.99998562
#> 214   male 106 0.99978563
#> 215 female 107 0.99999995
#> 216   male 107 0.99999686
#> 217 female 108 1.00000000
#> 218   male 108 0.99999999
#> 219 female 109 1.00000000
#> 220   male 109 1.00000000
#> 221 female 110 1.00000000
#> 222   male 110 1.00000000
#> 223 female 111 1.00000000
#> 224   male 111 1.00000000
#> 225 female 112 1.00000000
#> 226   male 112 1.00000000
#> 227 female 113 1.00000000
#> 228   male 113 1.00000000
#> 229 female 114 1.00000000
#> 230   male 114 1.00000000
#> 231 female 115 1.00000000
#> 232   male 115 1.00000000
#> 233 female 116 1.00000000
#> 234   male 116 1.00000000
#> 235 female 117 1.00000000
#> 236   male 117 1.00000000
#> 237 female 118 1.00000000
#> 238   male 118 1.00000000
#> 239 female 119 1.00000000
#> 240   male 119 1.00000000
#> 241 female 120 1.00000000
#> 242   male 120 1.00000000